% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries}
\alias{boundaries}
\title{Riding boundaries}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1600 rows and 3 columns.}
\usage{
boundaries
}
\description{
Boundaries for \link{ridings} for ridings that have  been in place since 2003. Derived from
the Open Canada Datasets for Federal Electoral Districts from
\href{https://open.canada.ca/data/en/dataset/78400aed-2370-4437-97ca-7563c21bacb1}{2003},
\href{https://open.canada.ca/data/en/dataset/10801c67-7f18-4ea1-bda7-8962abfc5578}{2013} and
\href{https://open.canada.ca/data/en/dataset/737be5ea-27cf-48a3-91d6-e835f11834b0}{2015}.
}
\keyword{datasets}
