% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/election-data-helpers.R
\name{election_dates}
\alias{election_dates}
\alias{election_years}
\alias{election_provinces}
\title{Choices for election years and provinces}
\usage{
election_dates()

election_years()

election_provinces()
}
\description{
Choices for election years and provinces
}
\examples{
election_dates()
election_years()
election_provinces()

}
