% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/election-data-helpers.R
\name{election_results}
\alias{election_results}
\title{Fetch election results in tidy format}
\usage{
election_results(
  years = election_years(),
  provinces = election_provinces(),
  results = "Elected"
)
}
\arguments{
\item{years}{One or more \code{\link[=election_years]{election_years()}} to include in the data}

\item{provinces}{One or more \code{\link[=election_provinces]{election_provinces()}} (or terretories)
include in the data. The order of \code{provinces} is kept in the data
by making the \code{province} variable a factor.}

\item{results}{One or both of "Elected" and/or "Defeated" to include
in the data}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} of a join between \link{results} and \link{ridings}.
}
\description{
Fetches election results in the format that you probably want for
plotting. For more complex operations, use joins of \link{results},
\link{ridings}, and/or \link{boundaries}. Note that in elections before
1968 there was more than one MP elected from several ridings.
}
\examples{
election_results(
  years = 2015:2019,
  provinces = c("Nova Scotia", "New Brunswick")
)

}
