% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{layout_province_grid}
\alias{layout_province_grid}
\title{Provincial grid layout}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10921 rows and 5 columns.}
\usage{
layout_province_grid
}
\description{
A layout that can be joined to \link{results} that works well facetted by
\code{province} with \code{geom_tile()}.
}
\keyword{datasets}
