% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_party.R
\docType{data}
\name{party_colour}
\alias{party_colour}
\alias{pal_party}
\alias{scale_colour_party}
\alias{scale_fill_party}
\alias{canadian_party_colours}
\alias{ScaleParty}
\title{Colours for Canadian Political Parties}
\format{An object of class \code{ScaleParty} (inherits from \code{ScaleDiscrete}, \code{Scale}, \code{ggproto}, \code{gg}) of length 2.}
\usage{
party_colour(
  party,
  party_colours = canadian_party_colours(),
  palette_extra = scales::hue_pal(),
  na.value = "grey50"
)

pal_party(
  party = character(0),
  party_colours = canadian_party_colours(),
  palette_extra = scales::hue_pal()
)

scale_colour_party(
  ...,
  party_colours = canadian_party_colours(),
  palette_extra = scales::hue_pal(),
  aesthetics = "colour"
)

scale_fill_party(
  ...,
  party_colours = canadian_party_colours(),
  palette_extra = scales::hue_pal(),
  aesthetics = "fill"
)

canadian_party_colours()

ScaleParty
}
\arguments{
\item{party}{A vector of party names like those seen in the \code{party} column of
\link{results}.}

\item{party_colours}{Known party colours, like those returned by
\code{\link[=canadian_party_colours]{canadian_party_colours()}}.}

\item{palette_extra}{A palette for assigning colours to rare parties like \code{\link[scales:hue_pal]{scales::hue_pal()}}.}

\item{na.value}{The colour to be used for \code{NA} values.}

\item{aesthetics, ...}{Passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
Colours for Canadian Political Parties
}
\examples{
party_colour(c(NA, "Liberal Party of Canada", "Conservative Party of Canada"))

library(ggplot2)

cols_df <- tibble::tibble(
  col = canadian_party_colours(),
  party = names(col),
  n = seq_along(col),
  x = n \%/\% 4,
  y = n \%\% 4
)

ggplot(cols_df, aes(x, y, col = party))  +
  geom_point(size = 10) +
  scale_colour_party()

ggplot(cols_df, aes(x, y, fill = party))  +
  geom_tile() +
  scale_fill_party()

}
\keyword{datasets}
