% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-election.R
\name{plot_election_map}
\alias{plot_election_map}
\alias{plot_election}
\title{Plot an election map}
\usage{
plot_election_map(years = 2006:2019, provinces = election_provinces())

plot_election(years = 2006:2019, provinces = election_provinces())
}
\arguments{
\item{years}{A vector of \code{\link[=election_years]{election_years()}} that will be included
in the plot. Riding geography is currently only available
for elections after 2006.}

\item{provinces}{One or more \code{\link[=election_provinces]{election_provinces()}} (or terretories)
include in the data. The order of \code{provinces} is kept in the data
by making the \code{province} variable a factor.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
\description{
Plot an election map
}
\examples{
plot_election(2019, "Ontario")
plot_election_map(2019, "Ontario")

}
