% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-election.R
\name{plot_votes}
\alias{plot_votes}
\alias{plot_seats}
\title{Plot the results of an election (bar chart)}
\usage{
plot_votes(
  years = 2006:2019,
  provinces = election_provinces(),
  x_var = "election_year",
  fill_var = "party",
  facet_var = NULL,
  facet_scales = "fixed",
  position = "stack",
  n_parties = 5
)

plot_seats(
  years = 2006:2019,
  provinces = election_provinces(),
  x_var = "election_year",
  fill_var = "party",
  facet_var = NULL,
  facet_scales = "fixed",
  position = "stack"
)
}
\arguments{
\item{years}{One or more \code{\link[=election_years]{election_years()}} to include in the data}

\item{provinces}{One or more \code{\link[=election_provinces]{election_provinces()}} (or terretories)
include in the data. The order of \code{provinces} is kept in the data
by making the \code{province} variable a factor.}

\item{x_var, fill_var}{A variable in \code{\link[=election_results]{election_results()}} to map to the \code{x} or \code{fill}
aesthetics, respectively}

\item{facet_var}{A variable to use with \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}}

\item{facet_scales}{Scales, most usefully "free" or "fixed".}

\item{position}{A position, most usefully "stack" or "fill".}

\item{n_parties}{The number of parties to include.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
\description{
Plots the number of votes or the number of seats as a bar chart. Use
\code{position = "fill"} to visualize vote proportions.
}
\examples{
plot_votes()
plot_seats()

}
