% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ridings}
\alias{ridings}
\title{Riding names and approximate geography}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1493 rows and 9 columns.}
\usage{
ridings
}
\description{
Derived from the Library of Parliament's
\href{https://lop.parl.ca/sites/ParlInfo/default/en_CA/}{ParlInfo site} for
\href{https://lop.parl.ca/sites/ParlInfo/default/en_CA/ElectionsRidings/Ridings}{ridings}.
The \code{riding_id} column refers to the ID used by the ParlInfo database.
}
\keyword{datasets}
