% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electionca-package.R
\docType{package}
\name{electionca-package}
\alias{electionca}
\alias{electionca-package}
\title{electionca: Canadian Elections Data}
\description{
Canadian elections data.
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (0000-0002-9415-4582)

}
\keyword{internal}
