% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{results}
\alias{results}
\title{Election Results}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 42096 rows and 7 columns.}
\usage{
results
}
\description{
Derived from the Library of Parliament's
\href{https://lop.parl.ca/sites/ParlInfo/default/en_CA/}{ParlInfo site} for
\href{https://lop.parl.ca/sites/ParlInfo/default/en_CA/ElectionsRidings/Elections}{election candidates and results}.
The \code{person_id} column refers to the ID used by the ParlInfo database.
}
\keyword{datasets}
